+++ pkg/bots/before_date_state.go
@@ -1,0 +1,4 @@
+package bots
+
+type BeforeDateState struct {
+}
